Make completion actions *never* ignore case
I often want to search using /, *, and friends in a case insensitive
manner, but it is seldom the case that if I'm halfway through a word
and I want to complete it that I want words that don't match in case
to be offered as completions.  This is i_CTRL-p and friends' behavior,
so I'm overriding it.  I might go back on this if I find this too
annoying, especially for things like plain text documents.