Add date of hovered over command to zsh_history status line
Zsh history files (if you have EXTENDED_HISTORY on) have the
timestamp of the entry included.  This is great unless you're
a human looking for a command run on a specific date or around
a specific time (this happens to me more often than I'd like to admit!).
This change adds a new "widget" to the status line when viewing zsh
history files that displays the human readable time for the current entry.