Experiment to stop using autochdir
...to enable plugins like Ack.vim and file finders to work
"properly".

I realized I would like to benefit from file finder plugins (among
others), but autochdir always got in the way.  When thinking about it a
bit, it dawned on me that I only make use of autochdir's behavior with a
few select commands (such as :edit).  After a bit of searching,
I stumbled upon a StackOverflow post (https://stackoverflow.com/questions/2170340/vim-e-starting-directory),
which recommending creating command mappings to automatically insert the
path to the current buffer, which would provide much of the same
functionality as autochdir, so I figured I'd give that a shot.