Only enable ALE for certain languages
I read a lot of code in various languages - ALE isn't super useful
when reading code, and it can be extremely annoying for certain
languages until you have things set up properly (eg. for C, Java,
and JavaScript).  So I figured it's better to explicitly opt-in for
languages that I like using with ALE