Fix folding for async functions
Vim's syntax facility has an optimization for keywords for fast
matching, but this causes tokens matching for keywords to be "consumed"
and unavailable for other match types, which means you can't use them
for things like folds.

So if JS folding is on, remove `async` from the list of keywords so that
we can fold based on its presence!